Make usage of coverage library consistent along the repo
