+++ book/src/02_development/01_tools.md
@@ -65,7 +65,7 @@ This tool compiles and instruments Solidity source files for use in EDR tests.
 
 Some integration tests require pre-compiled bytecode. To avoid adding `foundry-compilers` as a test dependency (which can interfere with other test crates that use solc during `cargo test --workspace`), we compile contracts ahead of time. It can also instrument source files with coverage probes, producing instrumented `.sol` files used by the `edr_solidity_tests` crate.
 
-By convention, compiled bytecode in EDR is kept in `data/deployed_bytecode/` as hex-encoded `.in` files. Integration tests load them via `include_str!` so they don't need solc at test time.
+By convention, compiled creation bytecode in EDR is kept in `data/creation_bytecode/` as hex-encoded `.bin` files (matching `solc --bin` output). Integration tests load them via `include_str!` so they don't need solc at test time.
 
 ### Compile with coverage instrumentation
 
@@ -88,20 +88,20 @@ cargo run -p edr_tool_solidity -- --instrument-only \
 
 ### Write bytecodes to disk
 
-Use `-o` to write `<ContractName>.in` files to a directory:
+Use `-o` to write `<ContractName>.bin` files to a directory:
 
 ```bash
 cargo run -p edr_tool_solidity -- --instrument \
-  -o data/deployed_bytecode \
+  -o data/creation_bytecode \
   data/contracts/test/CoverageTest.sol
 ```
 
 ### Compile with explicit imports
 
-Use `-i` to include additional source files needed by imports. For example, `increment.sol` is a pre-instrumented contract that already contains coverage probe calls and imports `coverage.sol` directly:
+Use `-i` to include additional source files needed by imports. For example, `Increment.sol` is a manually instrumented contract that imports `coverage.sol` directly:
 
 ```bash
-cargo run -p edr_tool_solidity -- data/contracts/increment.sol \
+cargo run -p edr_tool_solidity -- data/contracts/test/Increment.sol \
   -i data/contracts/coverage.sol
 ```
 
@@ -109,7 +109,7 @@ cargo run -p edr_tool_solidity -- data/contracts/increment.sol \
 
 Solidity contracts used by EDR tests live in `data/contracts/`. The coverage instrumentation library is at `data/contracts/coverage.sol`.
 
-When adding or modifying contracts, re-run the compile tool to regenerate the `.in` files and update the corresponding test code with any new function selectors shown in the tool output.
+When adding or modifying contracts, re-run the compile tool to regenerate the `.bin` files and update the corresponding test code with any new function selectors shown in the tool output.
 
 ## JS runner
