+++ crates/edr_provider/tests/integration/mod.rs
@@ -1,5 +1,4 @@
 mod coverage;
-mod coverage_scenarios;
 mod disable_balance_check;
 mod eip2537;
 mod eip4844;
